refactor SemanticMatcher class
I was having a hard time writing tests based on the example
in the project README.
So I've refactored the SemanticMatcher
class to understand how it works and begin building a
set of tests for the whole module.
SemanticMatcher was an implementation of algorithm 3
in the original research paper.
I started with it because of the 3 algorithm described in the
research paper it was the most isolated.